requirements.txt
torch==1.10.0
torchvision==0.11.1
qiskit==0.20.0
numpy==1.21.4
Pillow==8.4.0
matplotlib==3.4.3
